Begin supporting dongle channel changes.
This doesn't use a queue or anything - it just takes a single
frequency change that is done once the current sample has been read.
I'm attempting to not hold the lock during the rtlsdr work but in
tests the UI still hangs - I'm not yet sure why.

But this, with a quick hack to fm_sdl.c to directly call the dongle
frequency change routine, does seem to do the right thing..